Move `ExerciseQuery` to its own file
