+++ WorkoutTrack/Sources/WorkoutTrack/Exercises Cache/ExerciseStore.swift
@@ -12,15 +12,3 @@ public protocol ExerciseStore {
     func retrieve(by query: ExerciseQuery) async throws -> [CustomExercise]
     func delete(_ exercise: CustomExercise) async throws
 }
-
-public enum ExerciseQuery {
-    case all(sort: ExerciseSort?)
-    case byID(UUID, sort: ExerciseSort?)
-    case byName(String, sort: ExerciseSort?)
-    case byCategory(String, sort: ExerciseSort?)
-}
-
-public enum ExerciseSort {
-    case name(ascending: Bool)
-    case category(ascending: Bool)
-}
